backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pydantic==2.10.5
pydantic[email]==2.10.5
pydantic-settings==2.6.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.2
email-validator==2.2.0
